Move replicate_platinum and setup_fixtures into a control module
